// src/shared/api/generated/models/aiOperationListResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { AiOperationListItemDto } from "./aiOperationListItemDto";
import type { AiOperationPageInfoDto } from "./aiOperationPageInfoDto";

export interface AiOperationListResponseDto {
  items: AiOperationListItemDto[];
  pageInfo: AiOperationPageInfoDto;
}