// src/shared/api/generated/models/aiOperationProtectedAccessPageResponseDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */
import type { AiOperationProtectedAccessItemDto } from "./aiOperationProtectedAccessItemDto";
import type { AiOperationPageInfoDto } from "./aiOperationPageInfoDto";

export interface AiOperationProtectedAccessPageResponseDto {
  items: AiOperationProtectedAccessItemDto[];
  pageInfo: AiOperationPageInfoDto;
}